+++ src/app/src/main/kotlin/org/imperial/mrc/hint/controllers/LoginController.kt
@@ -7,16 +7,27 @@ import org.springframework.ui.set
 import org.springframework.web.bind.annotation.GetMapping
 import javax.servlet.http.HttpServletRequest
 import org.imperial.mrc.hint.AppProperties
+import org.imperial.mrc.hint.Translate
+import org.imperial.mrc.hint.security.oauth2.OAuth2AuthenticationRedirection
+import org.imperial.mrc.hint.security.oauth2.OAuth2State
 
 @Controller
-class LoginController(request: HttpServletRequest,
-                      private val session: Session,
-                      private val appProperties: AppProperties): Controllers(request)
+class LoginController(
+    private val request: HttpServletRequest,
+    private val session: Session,
+    private val translate: Translate,
+    appProperties: AppProperties,
+    oauth2State: OAuth2State
+) : OAuth2AuthenticationRedirection(appProperties, oauth2State)
 {
-
     @GetMapping("/login")
-    fun login(model: Model): String
+    fun login(model: Model): Any // Return type is string for formLogin, or ResponseEntity for OAuth2
     {
+        if (appProperties.oauth2LoginMethod)
+        {
+            return oauth2LoginRedirect()
+        }
+        println(request.getHeader("Accept-Language"))
         model["title"] = "Login"
         model["username"] = request.getParameter("username") ?: ""
         model["error"] = if (request.getParameter("error") == null)
@@ -25,11 +36,11 @@ class LoginController(request: HttpServletRequest,
         }
         else if (request.getParameter("error") == "SessionExpired")
         {
-            request.getParameter("message") ?: translateMessage("sessionExpiredLogin")
+            request.getParameter("message") ?: translate.key("sessionExpiredLogin")
         }
         else
         {
-            translateMessage("badUsernamePassword")
+            translate.key("badUsernamePassword")
         }
 
         val redirectTo = request.getParameter("redirectTo")