+++ src/app/src/test/kotlin/org/imperial/mrc/hint/unit/controllers/LoginControllerTests.kt
@@ -10,6 +10,13 @@ import org.junit.jupiter.api.Test
 import org.springframework.ui.ConcurrentModel
 import javax.servlet.http.HttpServletRequest
 import org.imperial.mrc.hint.ConfiguredAppProperties
+import org.imperial.mrc.hint.helpers.readPropsFromTempFile
+import org.imperial.mrc.hint.security.oauth2.OAuth2State
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import java.net.URI
 
 class LoginControllerTests
 {
@@ -18,7 +25,7 @@ class LoginControllerTests
     {
         val model = ConcurrentModel()
         val mockRequest = mock<HttpServletRequest>()
-        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties())
+        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties(), mock())
 
         val result = sut.login(model)
 
@@ -37,7 +44,7 @@ class LoginControllerTests
             on { getParameter("username") } doReturn "testUser"
             on { getParameter("error") } doReturn "CredentialsException"
         }
-        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties())
+        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties(), mock())
 
         val result = sut.login(model)
 
@@ -87,7 +94,7 @@ class LoginControllerTests
             on { getParameter("message") } doReturn "Some user message"
             on { getParameter("error") } doReturn "SessionExpired"
         }
-        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties())
+        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties(), mock())
 
         val result = sut.login(model)
 
@@ -104,7 +111,7 @@ class LoginControllerTests
         val mockRequest = mock<HttpServletRequest> {
             on { getParameter("error") } doReturn "SessionExpired"
         }
-        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties())
+        val sut = LoginController(mockRequest, mock(), ConfiguredAppProperties(), mock())
 
         val result = sut.login(model)
 
@@ -152,7 +159,7 @@ class LoginControllerTests
             on { getParameter("redirectTo") } doReturn "explore"
         }
         val mockSession = mock<Session>()
-        val sut = LoginController(mockRequest, mockSession, ConfiguredAppProperties())
+        val sut = LoginController(mockRequest, mockSession, ConfiguredAppProperties(), mock())
 
         val result = sut.login(model)
 
@@ -168,12 +175,39 @@ class LoginControllerTests
         val model = ConcurrentModel()
         val mockRequest = mock<HttpServletRequest>()
         val mockSession = mock<Session>()
-        val sut = LoginController(mockRequest, mockSession, ConfiguredAppProperties())
+        val sut = LoginController(mockRequest, mockSession, ConfiguredAppProperties(), mock())
 
         val result = sut.login(model)
 
         Assertions.assertThat(result).isEqualTo("login")
         Assertions.assertThat(model["appTitle"]).isEqualTo("Naomi")
         verify(mockSession).setRequestedUrl(null)
+    }
+
+    @Test
+    fun `login can redirect to auth0 tenant`()
+    {
+        val props = readPropsFromTempFile("oauth2_login_method=true")
+        val appProperties = ConfiguredAppProperties(props)
+
+        val model = ConcurrentModel()
+        val mockRequest = mock<HttpServletRequest>()
+        val mockStateCode = mock<OAuth2State> {
+            on { generateCode() } doReturn "xyz"
+        }
+        val sut = LoginController(mockRequest, mock(), appProperties, mockStateCode)
+
+        val result = sut.login(model) as ResponseEntity<*>
+
+        val httpHeader = HttpHeaders()
+        httpHeader.location = URI(
+            "https://fakeUrl/authorize?response_type=code&client_id=fakeId&" +
+                    "state=xyz&" +
+                    "scope=openid+profile+email+read:dataset&audience=naomi&" +
+                    "redirect_uri=http://localhost:8080/callback/oauth2Client"
+        )
+
+        assertEquals(result.statusCode, HttpStatus.SEE_OTHER)
+        assertEquals(result.headers.location, httpHeader.location)
     }
 }